change type of mode

+++ applicationLibrary/molecularDynamics/molecularDynamicsLibrary/MieFunctor.h
@@ -1356,7 +1356,7 @@ class MieFunctor
 
   ParticlePropertiesLibrary<SoAFloatPrecision, size_t> *_PPLibrary = nullptr;
 
-  static constexpr int mode = 0;
+  static constexpr uint8_t mode = 0;
   // sum of the potential energy, only calculated if calculateGlobals is true
   double _potentialEnergySum;
